+++ feature-pokedex/src/main/java/com/victor/feature_pokedex/presentation/ui/navigation/PokedexNavHost.kt
@@ -36,7 +36,7 @@ internal fun PokedexNavHost(
                 navArgument("pokemonId") { type = NavType.LongType }
             )
         ) {
-            val pokemonId = it.arguments?.getLong("pokemonId") ?: 0
+            val pokemonId = it.arguments?.getInt("pokemonId") ?: 0
             DetailsScreenBody(
                 navController = navController,
                 viewModel = viewModel,
